+++ src/ZendAuthenticationFactory.php
@@ -2,7 +2,7 @@
 /**
  * @see https://github.com/zendframework/zend-exprsesive-authentication-zendauthentication
  *     for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (http://www.zend.com)
  * @license https://github.com/zendframework/zend-exprsesive-authentication-zendauthentication/blob/master/LICENSE.md
  *     New BSD License
  */
@@ -10,15 +10,14 @@
 namespace Zend\Expressive\Authentication\ZendAuthentication;
 
 use Psr\Container\ContainerInterface;
+use Psr\Http\Message\ResponseInterface;
 use Zend\Authentication\AuthenticationService;
 use Zend\Expressive\Authentication\Exception;
-use Zend\Expressive\Authentication\ResponsePrototypeTrait;
-use Zend\Expressive\Authentication\UserRepositoryInterface;
+
+use function sprintf;
 
 class ZendAuthenticationFactory
 {
-    use ResponsePrototypeTrait;
-
     public function __invoke(ContainerInterface $container) : ZendAuthentication
     {
         $auth = $container->has(AuthenticationService::class)
@@ -43,7 +42,7 @@ class ZendAuthenticationFactory
         return new ZendAuthentication(
             $auth,
             $config,
-            $this->getResponsePrototype($container)
+            $container->get(ResponseInterface::class)
         );
     }
 }