+++ test/ZendAuthenticationTest.php
@@ -2,7 +2,7 @@
 /**
  * @see https://github.com/zendframework/zend-exprsesive-authentication-zendauthentication
  *     for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (http://www.zend.com)
  * @license https://github.com/zendframework/zend-exprsesive-authentication-zendauthentication/blob/master/LICENSE.md
  *     New BSD License
  */
@@ -10,25 +10,38 @@
 namespace ZendTest\Expressive\Authentication\Adapter;
 
 use PHPUnit\Framework\TestCase;
-use Prophecy\Argument;
-use Psr\Http\Message\ServerRequestInterface;
+use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 use Zend\Authentication\Adapter\AbstractAdapter;
 use Zend\Authentication\AuthenticationService;
 use Zend\Authentication\Result;
 use Zend\Expressive\Authentication\AuthenticationInterface;
 use Zend\Expressive\Authentication\UserInterface;
-use Zend\Expressive\Authentication\UserRepositoryInterface;
 use Zend\Expressive\Authentication\ZendAuthentication\ZendAuthentication;
 
 class ZendAuthenticationTest extends TestCase
 {
+    /** @var ServerRequestInterface|ObjectProphecy */
+    private $request;
+
+    /** @var AuthenticationService|ObjectProphecy */
+    private $authService;
+
+    /** @var UserInterface|ObjectProphecy */
+    private $authenticatedUser;
+
+    /** @var callable */
+    private $responseFactory;
+
     protected function setUp()
     {
         $this->request = $this->prophesize(ServerRequestInterface::class);
         $this->authService = $this->prophesize(AuthenticationService::class);
         $this->authenticatedUser = $this->prophesize(UserInterface::class);
-        $this->responsePrototype = $this->prophesize(ResponseInterface::class);
+        $this->responseFactory = function () {
+            return $this->prophesize(ResponseInterface::class)->reveal();
+        };
     }
 
     public function testConstructor()
@@ -36,7 +49,7 @@ class ZendAuthenticationTest extends TestCase
         $zendAuthentication = new ZendAuthentication(
             $this->authService->reveal(),
             [],
-            $this->responsePrototype->reveal()
+            $this->responseFactory
         );
         $this->assertInstanceOf(AuthenticationInterface::class, $zendAuthentication);
     }
@@ -50,7 +63,7 @@ class ZendAuthenticationTest extends TestCase
         $zendAuthentication = new ZendAuthentication(
             $this->authService->reveal(),
             [],
-            $this->responsePrototype->reveal()
+            $this->responseFactory
         );
         $result = $zendAuthentication->authenticate($this->request->reveal());
         $this->assertInstanceOf(UserInterface::class, $result);
@@ -64,7 +77,7 @@ class ZendAuthenticationTest extends TestCase
         $zendAuthentication = new ZendAuthentication(
             $this->authService->reveal(),
             [],
-            $this->responsePrototype->reveal()
+            $this->responseFactory
         );
         $this->assertNull($zendAuthentication->authenticate($this->request->reveal()));
     }
@@ -77,7 +90,7 @@ class ZendAuthenticationTest extends TestCase
         $zendAuthentication = new ZendAuthentication(
             $this->authService->reveal(),
             [],
-            $this->responsePrototype->reveal()
+            $this->responseFactory
         );
         $this->assertNull($zendAuthentication->authenticate($this->request->reveal()));
     }
@@ -106,7 +119,7 @@ class ZendAuthenticationTest extends TestCase
         $zendAuthentication = new ZendAuthentication(
             $this->authService->reveal(),
             [],
-            $this->responsePrototype->reveal()
+            $this->responseFactory
         );
         $this->assertNull($zendAuthentication->authenticate($this->request->reveal()));
     }
@@ -136,7 +149,7 @@ class ZendAuthenticationTest extends TestCase
         $zendAuthentication = new ZendAuthentication(
             $this->authService->reveal(),
             [],
-            $this->responsePrototype->reveal()
+            $this->responseFactory
         );
         $result = $zendAuthentication->authenticate($this->request->reveal());
         $this->assertInstanceOf(UserInterface::class, $result);